Move shell completion scripts to pulseaudio-utils
The shell completion scripts support commands from both the pulseaudio
package and pulseaudio-utils package, which may be installed without
pulseaudio (e.g. for use with PipeWire configured for PulseAudio
compatibility).  Ship the completions in pulseaudio-utils to cover this
case.  Note that the completion for the pulseaudio command will be
available if the pulseaudio package is installed, since it depends on
pulseaudio-utils.

Also move the Lintian override for script-not-executable and add one for
bash-completion-with-hashbang.  I'm in favor of fixing these, rather
than overriding them, but not sufficiently to pursue it right now.

Closes: #803329

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>